+++ test/integration/model/findAll/groupedLimit.test.js
@@ -2,6 +2,7 @@
 
 /* jshint -W030 */
 var chai = require('chai')
+  , sinon = require('sinon')
   , expect = chai.expect
   , Support = require(__dirname + '/../../support')
   , Sequelize = Support.Sequelize
@@ -14,6 +15,18 @@ if (current.dialect.supports['UNION ALL']) {
   describe(Support.getTestDialectTeaser('Model'), function() {
     describe('findAll', function () {
       describe('groupedLimit', function () {
+        before(function () {
+          this.clock = sinon.useFakeTimers();
+        });
+
+        afterEach(function () {
+          this.clock.reset();
+        });
+
+        after(function () {
+          this.clock.restore();
+        });
+
         beforeEach(function () {
           this.User = this.sequelize.define('user', {
             age: Sequelize.INTEGER
@@ -22,8 +35,19 @@ if (current.dialect.supports['UNION ALL']) {
             title: DataTypes.STRING
           });
           this.Task = this.sequelize.define('task');
+
+          this.ProjectUserParanoid = this.sequelize.define('project_user_paranoid', {}, {
+            timestamps: true,
+            paranoid: true,
+            createdAt: false,
+            updatedAt: false
+          });
+
           this.User.Projects = this.User.belongsToMany(this.Project, {through: 'project_user' });
           this.Project.belongsToMany(this.User, {as: 'members', through: 'project_user' });
+
+          this.User.ParanoidProjects = this.User.belongsToMany(this.Project, {through: this.ProjectUserParanoid});
+          this.Project.belongsToMany(this.User, {as: 'paranoidMembers', through: this.ProjectUserParanoid});
 
           this.User.Tasks = this.User.hasMany(this.Task);
 
@@ -40,6 +64,8 @@ if (current.dialect.supports['UNION ALL']) {
               return Promise.join(
                 projects[0].setMembers(users.slice(0, 4)),
                 projects[1].setMembers(users.slice(2)),
+                projects[0].setParanoidMembers(users.slice(0, 4)),
+                projects[1].setParanoidMembers(users.slice(2)),
                 users[2].setTasks(tasks)
               );
             });
@@ -133,6 +159,55 @@ if (current.dialect.supports['UNION ALL']) {
                */
               expect(users).to.have.length(5);
               expect(users.map(u => u.get('id'))).to.deep.equal([1, 3, 5, 7, 4]);
+            });
+          });
+
+          it('works with paranoid junction models', function() {
+            return this.User.findAll({
+              attributes: ['id'],
+              groupedLimit: {
+                limit: 3,
+                on: this.User.ParanoidProjects,
+                values: this.projects.map((item) => item.get('id'))
+              },
+              order: [
+                Sequelize.fn('ABS', Sequelize.col('age')),
+                ['id', 'DESC']
+              ],
+              include: [this.User.Tasks]
+            }).then((users) => {
+              /*
+                project1 - 1, 3, 4
+                project2 - 3, 5, 7
+               */
+              expect(users).to.have.length(5);
+              expect(users.map((u) => u.get('id'))).to.deep.equal([1, 3, 5, 7, 4]);
+
+              return Sequelize.Promise.join(
+                this.projects[0].setParanoidMembers(users.slice(0, 2)),
+                this.projects[1].setParanoidMembers(users.slice(4))
+              );
+            }).then(() => {
+              return this.User.findAll({
+                attributes: ['id'],
+                groupedLimit: {
+                  limit: 3,
+                  on: this.User.ParanoidProjects,
+                  values: this.projects.map((item) => item.get('id'))
+                },
+                order: [
+                  Sequelize.fn('ABS', Sequelize.col('age')),
+                  ['id', 'DESC']
+                ],
+                include: [this.User.Tasks]
+              });
+            }).then((users) => {
+              /*
+                project1 - 1, 3
+                project2 - 4
+               */
+              expect(users).to.have.length(3);
+              expect(users.map((u) => u.get('id'))).to.deep.equal([1, 3, 4]);
             });
           });
         });